+++ tests/TestCase/Model/Table/PagesCategoriesTableTest.php
@@ -115,6 +115,6 @@ class PagesCategoriesTableTest extends TableTestCase
         $sql = $query->sql();
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertInstanceOf(FrozenTime::class, $query->getValueBinder()->bindings()[':c1']['value']);
-        $this->assertSqlEndsWith('FROM pages_categories Categories INNER JOIN pages Pages ON (Pages.active = :c0 AND Pages.created <= :c1 AND Categories.id = Pages.category_id)', $sql);
+        $this->assertStringEndsWith('FROM pages_categories Categories INNER JOIN pages Pages ON (Pages.active = :c0 AND Pages.created <= :c1 AND Categories.id = Pages.category_id)', $sql);
     }
 }